+++ test/server_statistics_test.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:test/test.dart';
 
 import 'package:open62541/open62541.dart';
@@ -31,8 +29,8 @@ void main() {
     late int port;
     late Server server;
 
-    setUp(() {
-      port = Random().nextInt(10000) + 4840;
+    setUp(() async {
+      port = await freeTcpPort();
       server = setupServer(port);
       addBasicVariables(server);
     });